# openapi.yml
openapi: 3.0.0
info:
  title: Botapeer API
  description: Botapeer API
  version: 1.0.0
servers:
  - url: "{server}/api"
    variables:
      server:
        default: http://localhost:8081
        enum:
          - http://localhost:8081
          - https://api.botapeer.com
paths:
  '/users':
    get:
      operationId: getUsersOrGetUserByName
      tags:
        - User
      description: User一覧もしくは名前からユーザーを取得
      parameters:
        - in: query
          name: username
          description: ユーザー名
          required: false
          schema:
            type: string
            example: taro
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '404':
          description: 404(User Not Found)
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: 405(Validation exception)
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ErrorResponse'
  '/users/{userId}':
    get:
      operationId: findUserById
      tags:
        - User
      description: 任意のUserをuserIdから取得
      parameters:
        - in: path
          name: userId
          description: ユーザーID
          required: true
          schema:
            type: string
            example: 1
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: 404(User Not Found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: 405(Validation exception)
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ErrorResponse'
    patch:
      operationId: updateUser
      tags:
        - User
      description: 任意のUserをuserIdから更新
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          description: ユーザーID
          required: true
          schema:
            type: string
            example: 1
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              properties:
                formData:
                  $ref: '#/components/schemas/UpdateUserFormData'
                profileImage:
                  type: string
                  format: binary
                coverImage:
                  type: string
                  format: binary
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: 401(Unauthorized)
        '405':
          description: 405(Validation exception)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteUser
      tags:
        - User
      description: 任意のUserをuserIdから削除
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          description: ユーザーID
          required: true
          schema:
            type: string
            example: 1
      responses:
        '200':
          description: ok
        '401':
          description: 401(Unauthorized)
        '405':
          description: 405(Validation exception)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
  '/users/plant_records/{plantRecordId}':
    get:
      operationId: findUserByPlantRecordId
      tags:
        - User
      description: 任意のUserをPlantRecordIdから取得
      parameters:
        - in: path
          name: plantRecordId
          description: 生育記録ID
          required: true
          schema:
            type: string
            example: 1
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: 404(User Not Found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: 405(Validation exception)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
...
components:
  schemas:
    UpdateUserFormData:
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 15
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 255
        description:
          type: string
      required:
        - name
        - email
    UserResponse:
      properties:
        id:
          type: integer
        name:
          type: string
          minLength: 1
          maxLength: 15
        email:
          type: string
          format: email
        status:
          type: integer
        description:
          type: string
        profileImage:
          type: string
        coverImage:
          type: string
...
    ErrorResponse:
      properties:
        errors:
          $ref: '#/components/schemas/Error'
    Error:
        type: array
        items:
          properties:
            code:
              type: integer
            message:
              type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT